requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==2.0.27
aiosqlite==0.19.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0 
email-validator==2.1.0
secure-smtplib==0.1.1
pydantic==2.5.3
gunicorn==21.2.0
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
httpx==0.26.0
jsonschema==4.20.0
openai==1.6.1
